Move RemoteHaystackComponent from deployments
